add check for size, check for price more then $200
