+++ tests/test_purchase_products_with_discount.py
@@ -2,6 +2,8 @@ import allure
 from selene import browser
 from selene.support.conditions import have
 
+from pages.cart_page import CartPage
+from pages.erin_recommends_page import ErinRecommendsPage
 from pages.main_page import MainPage
 
 
@@ -14,16 +16,24 @@ class TestPurchaseOfGoodsWithDiscount:
             page = MainPage(browser=browser)
             page.open_page()
             page.is_loaded()
+            page.scroll_to_hot_sellers()
         with (allure.step("Add to products item to card until sub total > 200")):
             products = page.find_products()
-            page.scroll_to_hot_sellers()
             size = len(products)
+            if size == 0:
+                page = ErinRecommendsPage(browser=browser)
+                page.open_page()
+                size = len(page.find_products())
+                print(size)
             count = 0
             for i in range(0, size):
                 page.add_product_to_cart(products[i])
                 count += 1
                 page.find_counter_number().should(have.text(str(count)))
+                if page.get_subtotal() > 200:
+                    break
             page.find_counter_number().should(have.text(str(count)))
         with (allure.step("Go to checkout card with products item")):
-            card = page.goto_card_page()
+            card = CartPage(browser).open_page()
+            card.goto_card_page()
             print(card.get_cart_totals())
